requirements.txt
asgiref==3.8.1
Django==4.1.13
django-cors-headers==4.4.0
django-filter==24.2
django-redis==5.4.0
djangorestframework==3.15.2
djongo==1.3.6
dnspython==2.6.1
Markdown==3.6
pymongo==3.12.3
pytz==2024.1
redis==5.0.6
sqlparse==0.2.4
django-environ